and now added more files

+++ imu/lsm6/LSM6.h
@@ -90,8 +90,8 @@ class LSM6
     vector<int16_t> a_raw; // raw accelerometer readings
     vector<int16_t> g_raw; // raw gyro readings
     
-    vector<float> a; // accelerometer readings (g)
-    vector<float> g; // gyro readings (deg/s)
+    vector<int16_t> a; // accelerometer readings (milli-g)
+    vector<int16_t> g; // gyro readings (10 * deg/s)
 
     uint8_t last_status; // status of last I2C transmission
